Split out cli.py.
