+++ tox.ini
@@ -16,6 +16,7 @@ commands = flake8
 select = C,E,F,W,B,B9
 ignore = E221
 max-complexity = 7
+max-line-length = 80
 
 
 [testenv:lint]
